feat: add syncing and id helper
